moved Transaction back inside plunger.py and added typing
